+++ tools/fixWakeLock/tool.js
@@ -6,7 +6,7 @@
         name:'fixWakeLock',
         category: 'general',
         type:'single',
-        possibleValues: [true,false],
+        possibleValues: [false,true],
         defaultValue: true,
         advanced: true,
         hidden: true